Create LICENSE
I have created this license to ensure that I am able to receive credit for my work. With this, I allow total public use of this code, whether they be backyard astronomers, educators, or software developers. If VaRSL is to be used in any software--astronomy-based or otherwise--I ask to be credited as per the license above. Same goes for any research conducted using my code with the only exception being that my code just so happens to be in an astronomy software the researcher was using, assuming the software has credited me already for using my code.